FIX hp fortify
6.1.3 Type Mismatch: Negative to Unsigned
6.1.4 Type Mismatch: Signed to Unsigned

+++ ZXingObjC/client/result/ZXExpandedProductParsedResult.m
@@ -95,7 +95,7 @@ NSString * const ZX_POUND = @"LB";
 }
 
 - (NSUInteger)hash {
-  int hash = 0;
+  NSUInteger hash = 0;
   hash ^= [self.productID hash];
   hash ^= [self.sscc hash];
   hash ^= [self.lotNumber hash];
